requirements.txt
llama_index
scrapy
cssutils
pyvis
Ipython
langchain
pypdf
pymupdf
chainlit
torch 
transformers 
python-pptx 
Pillow
uptrain
gradio
llama-index-core
haystack-ai
streamlit
python-docx